app/src/main/java/com/app/marsrover/data/repository/MarsRoverManifestRepo.kt
package com.app.marsrover.data.repository

import com.app.marsrover.data.model.RoverManifestUiState
import com.app.marsrover.data.converters.toUiModel
import com.app.marsrover.service.MarsRoverManifestService
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flow
import javax.inject.Inject

class MarsRoverManifestRepo @Inject constructor(
    private val marsRoverManifestService: MarsRoverManifestService
) {
    fun getMarsRoverManifest(roverName: String): Flow<RoverManifestUiState> = flow {
        try {
            emit(
                toUiModel(
                    marsRoverManifestService.getMarsRoverManifest(
                        roverName
                    )
                )
            )
        } catch (throwable: Throwable) {
            emit(RoverManifestUiState.Error)
        }

    }
}